+++ src/MainRoute.js
@@ -4,11 +4,9 @@ import { connect } from "unistore/react";
 import { actions } from "./store";
 import App from "./pages/welcomepage/welcome";
 import Home from "./pages/homepage/home";
-import Location from "./pages/servicepage/location";
-import Order from "./pages/orderpage/order";
+import Location from "./pages/orderpage/order";
+import Order from "./pages/orderhistory/orderhistory";
 import OrderDetails from "./pages/orderdetails/orderdetails";
-import Time from "./component/time/time";
-import Upload from "./component/time/foto";
 import NotFound from "./pages/notfound/NotFound";
 import Footer from "./component/footer";
 import Header from "./component/header";
@@ -17,6 +15,7 @@ import RewardPage from "./pages/rewardpage/reward";
 import Profile from "./pages/profile/profile";
 import Register from "./validcoba";
 import Help from "./pages/helppage/helppage";
+import ListSampah from "./pages/trashpage/trashpage";
 
 class MainRoute extends React.Component {
   render() {
@@ -25,8 +24,8 @@ class MainRoute extends React.Component {
         <Switch>
           <Route exact path="/" component={App} />
           <Route exact path="/home" component={Home} />
-          <Route path="/service" component={Location} />
-          <Route path="/order" component={Order} />
+          <Route path="/order" component={Location} />
+          <Route path="/orderhistory" component={Order} />
           <Route path="/orderdetails" component={OrderDetails} />
           <Route path="/reward" component={RewardPage} />
           <Route path="/profile" component={Profile} />
@@ -34,6 +33,7 @@ class MainRoute extends React.Component {
           <Route path="/header" component={Header} />
           <Route path="/register" component={Register} />
           <Route path="/help" component={Help} />
+          <Route path="/listsampah" component={ListSampah} />
           <Route component={NotFound} />
         </Switch>
       </Router>
