+++ src/pages/orderhistory/order.css
@@ -1,0 +1,5 @@
+@import url("https://fonts.googleapis.com/css?family=Muli:300i,400,600,600i,700,800,800i,900&display=swap");
+
+.font {
+  font-family: "Muli", sans-serif;
+}
